X11: Keep the menubar inside the widgetbox window in toplevel mode
For now the appmenu protocol does not make it possible to associate a
menubar with all application windows. This means in top level mode you
can only reach the menubar when the widgetbox window is active. Since
this is quite inconvenient, better not use the native menubar in this
configuration and keep the menubar in the widgetbox window.

Merge-request: 1229
Reviewed-by: denis

(cherry-picked from commit a45398677309ab905b69c599177b4c2951292b04)

Change-Id: Ib2994551558d87db84874f84f279be5e766cde15
Reviewed-on: http://codereview.qt-project.org/4602
Reviewed-by: Qt Sanity Bot <qt_sanity_bot@ovi.com>
Reviewed-by: Oswald Buddenhagen <oswald.buddenhagen@nokia.com>