Remove dependency on designer from linguist and headers autotest.
When compiling with shadow build option the Qt sources are not
available. The QtDesigner library requires direct access to these
sources to compile. Hence if we remove the dependency on designer
(which is not required) we can compile successfully with the
shadow build configure option.

Change-Id: I36fa408a32a354d54a9b47aed8a7cba447cf4129
Sanity-Review: Qt Sanity Bot <qt_sanity_bot@ovi.com>
Reviewed-by: Oswald Buddenhagen <oswald.buddenhagen@nokia.com>