Continue processing the qhp file on missing attributes
qhelpgenerator would bail out with an error if a keyword would contain
a missing or empty attribute. This can e.g. happen on a missing
\relates in the documentation.

This makes the whole help generation rather fragile, as CI doesn't stop
these kind of errors from making it into the code base. So rather then
failing to generate the docs simply skip the one keyword that the
helpgenerator can't place into a context.

Change-Id: I265dc75ba9b29c868d5b9cf8a9aae5ce880374b1
Reviewed-by: Casper van Donderen <casper.vandonderen@nokia.com>