Enable exceptions code, makes Assistant far more stable.
Task-number: QTBUG-28592

The reason to enable exceptions (always) is that we build the
full text engine with exceptions as well, which throws erratic.
Once it does that the database it creates is in a unknow state,
which leads to new exceptions on next Assistant start. So it will
never start again until someone removes the broken index by hand.

Change-Id: Ic24a9e28bddf4657753f9b670c38ecda527d7af4
Reviewed-by: Friedemann Kleint <Friedemann.Kleint@digia.com>
Reviewed-by: Joerg Bornemann <joerg.bornemann@digia.com>
Reviewed-by: Karsten Heimrich <karsten.heimrich@digia.com>