Move functions from an anonymous namespace in a header
Clang doesn't like functions declared in an anonymous namespace in a
header and then later not used. It treats those are static functions,
and the -Wunused-functions warning is printed for them.

Instead, declare them inline in a regular namespace and use it.

qclucene_global_p.h:106:19: error: unused function 'QStringToTChar' [-Werror,-Wunused-function]
    inline TCHAR* QStringToTChar(const QString &str)
                  ^

Change-Id: I94696d40693558ef105281c8ec6e30159c47844d
Reviewed-by: Karsten Heimrich <karsten.heimrich@digia.com>
Reviewed-by: Kai Koehne <kai.koehne@digia.com>