Try harder in matching the method signature
In cases where the Propertiesdialog has no specific Implementation
for a dbus data type, it creates a regular QLineEdit. In such cases,
qdbusviewer then tried to call the method with the arg set to a string
instead of the real datatype (for example uint64), which usually does not work.

We can be more clever and try to implicitely convert the user input
to the right datatype, which makes the invocation succeed.

Change-Id: Idbb5c667496e14922efb7d195e11fafbb880c61c
Reviewed-by: Thiago Macieira <thiago.macieira@intel.com>