make lupdate a host tool
lupdate makes no sense on x-built targets at all.
though one may also argue whether it makes sense to actually bootstrap
it - while its use may be supported by the build system, it's typically
not part of regular builds, but rather used shortly before running
linguist - for which a host build of qt is necessary anyway.
one argument for bootstrapping is that it uses the qmldevtools library,
which logically should be bootstrapped (the only other current user is
qmlmin, which definitely *is* used in regular builds and thus needs
bootstrapping).

Change-Id: Ia3ddd901fe2b0c2b7b77af7f1ebc66d757fc7783
Reviewed-by: hjk <hjk121@nokiamail.com>
Reviewed-by: Oswald Buddenhagen <oswald.buddenhagen@digia.com>