Fix assistant and designer builds without printer.
The actual code using printer related classes is already wrapped with
QT_NO_PRINTER macros, but the include statements were not.

The reason why include statement probably were not wrapped with
QT_NO_PRINTER is that in Qt4 all headers were synchronized to include
directory. Now with Qt5 modularization headers for modules which
are disabled are not anymore synchronized to include directory so
they need to be wrapped with same ifdefs.

Change-Id: I0efd1342cd18860b5fed63b248812e9dced8f2b3
Reviewed-by: Oswald Buddenhagen <oswald.buddenhagen@digia.com>