androiddeployqt: Fix overriding auto-detected deployment
The platform plugin is set as part of the whole autodetection
step. When this step is not run because deployment is explicitly
set in the .pro file,  we should still detect the platform plugin.

The way we do this is by checking which platform plugin is
included in the deployment settings and using this.

Change-Id: I189972a34f03b9059db84bc41855ac662071ca11
Reviewed-by: Paul Olav Tvete <paul.tvete@digia.com>