Android: Remove non-host tools
Command line applications for target architecture does not work on
Android, since apps there are Java applications which call into a
shared library. Thus, these command line tools which are not built for
host architecture are not usable in a SUBDIRS build. They need to be
built and deployed separately in apk packages.

Change-Id: Ia7322ee143a1b73ac40459f078e9f51a4e29a3ab
Reviewed-by: BogDan Vatra <bogdan@kde.org>