androiddeployqt: Remove support for QMAKE_SUPPORTED_ORIENTATIONS
This was created for the iOS port and intended to be cross-platform,
but since then they got cold feet and never added the support.
It makes no sense to only support this on Android. We need to
remove it again and hold off until we can find a proper solution.
Editing the AndroidManifest.xml is the non-cross-platform
solution for this.

Task-number: QTBUG-34166
Change-Id: Ide8abb159266bcb2af03667ecf8d92da127960b1
Reviewed-by: BogDan Vatra <bogdan@kde.org>