Extend androiddeployqt features
Makes it possible to specify a set of runtime dependencies to be bundled
with the application to be used afterwards.
To do so, adds a new entry in the json file named "android-extra-plugins"
that will specify those dependencies in a list.
It will put the text files within the assets directory and the libraries
and plugins in the architecture's directory in libs/.

Change-Id: I2aaa4886252743bd333d8d18404fd41e85bbd8b5
Reviewed-by: Eskil Abrahamsen Blomfeldt <eskil.abrahamsen-blomfeldt@digia.com>