Add font size storage/unified handling in Linguists editors
Add a unified font size handling in Linguists editors by assigning
Ctrl-+/Ctrl--/Ctrl-Wheel to inc/dec font size of ALL editors. The
current font size is also stored in the configuration and set at next
startup. This is done to make it simpler to use for translators.
If the configuration does not have a font size set it will set the
default system size.

Change-Id: I941f07807653360f8518938380f2c6a2d6f08b63
Reviewed-by: Oswald Buddenhagen <oswald.buddenhagen@digia.com>