Introducing winrtrunner
winrtrunner is a console application for launching WinRT Qt packages. It
handles:
- installing (registering) & removing the package
- launching & terminating the app
- fetching & dispaying test output for test cases

It is designed to compile everywhere, as the platform-specific backends
are left out where not supported. Currently it only has a backend for
Appx packages, so MSVC2012/2013 is required. It may support other
backends over time, such as Windows Phone and remote Appx, as well as
any host SDK that gains support for deploying to these platforms.

Done-with: Maurice Kalinowski <maurice.kalinowski@digia.com>
Change-Id: I424c228435d8eb4608a1d1854106b9df69ca5f11
Reviewed-by: Friedemann Kleint <Friedemann.Kleint@digia.com>
Reviewed-by: Oliver Wolff <oliver.wolff@digia.com>