qtd3dservice: Check that CoreCon is initialized with exists()
524375b attempted to fix multithreaded access with a mutex, but that
was not enough to handle access before the server was fully constructed.
This partially reverts 524375b and puts the initialization code into
a separate function which can be checked to make sure the server is
fully constructed before proceeding.

The Q_DECLARE_PRIVATE macro is added for clarity and const-correctness.

Change-Id: I0e87ee32e1b409f18b3c5fc4aab22d3c8a85cef8
Reviewed-by: Oliver Wolff <oliver.wolff@digia.com>