Long live qtdiag!
qtdiag is a command line tool that prints information about the Qt
version and the graphics configuration. Its output can be used
for CI logs and customer bug reports.

The code originally was in the manual test qpainfo and
has been turned into a function returning a QString which
can potentially be used in GUI applications as well.

Change-Id: Ic19b6064307eefa67ced7c551a09140b985a89cf
Reviewed-by: Laszlo Agocs <laszlo.agocs@digia.com>
Reviewed-by: Lars Knoll <lars.knoll@digia.com>