Add more visible/accessible font zooming in Linguist
Makes font zooming more visible by adding it to the view-menu.
In addition to Ctrl-+/- or Ctrl-ScrollWheel for controlling
font size it is now possible to control it by using the two
new menu items, increase and decrease, in the View->Zoom menu.

The font size can also be reset by using Ctrl-0 or a menu-item.

Task-number: QTBUG-37561

Change-Id: Ie3e61a138d007ea5435ef67e111aaa845f126081
Reviewed-by: Friedemann Kleint <Friedemann.Kleint@digia.com>
Reviewed-by: Oswald Buddenhagen <oswald.buddenhagen@digia.com>