androiddeployqt: Support bundling library projects
Third party library projects can easily be added to the Android
package source directory and added to the project using
project.properties, but "android update" needs to be run on it
to produce some local build files before it can be included in the
build. We can do this for people so that they can use third party
library projects without having to put build files in their source
directory.

[ChangeLog][Android][Deployment] Detect linked library projects
in the package source directory and automatically run "android update"
on these.

Task-number: QTBUG-38322
Change-Id: I0bf9b8708a7ec28ed5429379215b12a65d05416a
Reviewed-by: Christian Stromme <christian.stromme@digia.com>