Add option for skipping use of private API.
The qsqlodbc and qsqlpsql are known to cause rejections
from the Mac App store due to use of private/deprecated
API.

The plugins are deployed whenever the SQL module is used.
Due to this corse granularity of plugin selection apps
may find themselves rejected for plugins which they
are not actually using.

Make macdeployqt print a warning when sqlodbc or sqlpsql
is deployed. Add "-appstore-compliant" which will
skip deployment of the plugins.

Task-number: QTBUG-37835
Task-number: QTBUG-38607
Change-Id: I7325156ffaf228a97d7ceeb12f329b3f10db4ff2
Reviewed-by: Gabriel de Dietrich <gabriel.dedietrich@theqtcompany.com>