androiddeployqt: Remove warning about missing dependency
In 08672adb97c0707c513850a27aaec59327874260 in qtbase, the
MODULE_PLUGIN_TYPES of QtGui was expanded to contain some
more plugin types which are not built on Android, causing
warnings during every deployment. Since MODULE_PLUGIN_TYPES
is common to all platforms we should just ignore it if it
specifies plugins that do not exist.

[ChangeLog][androiddeployqt] Remove warning about missing
dependency.

Task-number: QTBUG-38651
Change-Id: I57a2c1d79b859981b3adf2772be4f0cd8a98be12
Reviewed-by: Caroline Chao <caroline.chao@digia.com>
Reviewed-by: Paul Olav Tvete <paul.tvete@digia.com>