macdeployqt: don't use hardcoded lib/ when searching for dylib
Currently when parsing otool output for dylibs, "lib/" is appended to
the path. However dylibs might not be always found in a lib folder like
e.g. the mysql client libraries from macports. This patch aims to remove
that hardcoded path and use the data from the currently parsed otool
line.

[ChangeLog][OS X][macdeployt] Fixed a bug where dynamic libraries where
searched in the wrong directory

Task-number: QTBUG-21913
Change-Id: I992d15bc48ad827185216852108b2d0a5633043e
Reviewed-by: Frederik Gladhorn <frederik.gladhorn@digia.com>
Reviewed-by: Jake Petroules <jake.petroules@petroules.com>