androiddeployqt: Refactor findFilesRecursively()
Currently, androiddeployqt expects all the detected dependencies
to reside inside the Qt installation directory. Once we add support
for the qmlimportscanner, however, they can reside in any of
the qml import paths. The refactoring prepares the code for this,
separating the relative path (the path of the file when it is
installed on the device) and the absolute path (the source file).

Task-number: QTBUG-34175
Change-Id: I5173034a02e29f88f8ca6f2b61c0caea6ba71778
Reviewed-by: BogDan Vatra <bogdan@kde.org>