androiddeployqt: Force Qt Gui dependency for all apps
Currently, the only way to launch an Android application is through
the platform plugin, which of course depends on Qt Gui. This causes
problems running corelib autotests, since they have QT = core in
the .pro file and thus will not deploy or load the platform plugin.

At some point, we may have a way to launch pure corelib-applications
on Android (like services), but until then, we need to hardcore the
dependency on Qt Gui to make sure these autotests run properly.

Change-Id: Ica794af972a06ec021c1fa28333866296eab96e0
Reviewed-by: BogDan Vatra <bogdan@kde.org>