Zoom the documentation according to the system DPI by default
QtWebKit follows the web standard use of 96 DPI for font sizes and
this is a good compromise for the web, but this would require each
Assistant user to zoom the text manually without the benefit of
keeping existing content working (since we don't show any external
content in Assistant that could break with a higher DPI).

HiDPI on OS X already handles the issue by scaling the whole page by
two, but since we don't completely support DPI awareness on Windows,
Assistant assets and text will be scaled but the documentation
itself will be microscopic.

Use the zoomFactor to adjust the DPI according to system settings
by dividing QScreen::logicalDotsPerInch() by 96.

Task-number: QTBUG-41076
Change-Id: I5644d0b383fcaf04bed156207afca9b322ea18a6
Reviewed-by: Friedemann Kleint <Friedemann.Kleint@digia.com>
Reviewed-by: Allan Sandfeld Jensen <allan.jensen@digia.com>