uic: don't use QStringLiteral in comparisons
For QLatin1String, operator== is overloaded, even for QStringRef,
so comparing to a latin-1 (C) string literal is efficient,
since strlen() is comparatively fast.

OTOH, QStringLiteral, when not using RVO, litters the code with
QString dtor calls, which are not inline. Worse, absent lambdas,
it even allocates memory.

So, just compare using QLatin1String instead.

Change-Id: Ice92d40342cb3939e8697d3f778f4421f5f967a1
Reviewed-by: Thiago Macieira <thiago.macieira@intel.com>