Assistant: Fix high CPU usage if a parent process finishes
If Assistant has been started in the remote control mode in Linux
and it's parent process finishes, then Assistant increases CPU load
to 100% and continuously raises it's window.
It is because QSocketNotifier for stdin continuously emits
activated() signal although there is no data to read.
So disable the notifier after receiving EOF.

Change-Id: I1b01108e1b97bc4fd54411b16489082be7f47bf4
Reviewed-by: Karsten Heimrich <karsten.heimrich@theqtcompany.com>
Reviewed-by: Christian Kandeler <christian.kandeler@theqtcompany.com>