androiddeployqt: Fix dependency order for XML dependencies
Although this bug was there in Qt 5.3 as well, something has changed
in our build system that causes readelf to return libraries in a
different order than before. In particular, the QtMultimedia dependency
is now listed before QtQuick if it's listed first in the QT variable in
the .pro file. This caused the following bug:

When we read the manual dependencies from the XML specification,
then we assume the dependencies of these are already met (otherwise
they should not be added). Therefore, we need to load these
dependencies last, not intermingled with the actual direct
dependencies of the application.

What would happen is that Qt5Multimedia came before Qt5Quick, and
triggered adding Qt5MultimediaQuick_p before Qt5Quick. This of course
depends on Qt5Quick so it needs to be added after it (which is why
it is only added if the application itself depends on Qt Quick).

Change-Id: Ifa97d85645cff49ebf65d4ce3fc2101506346695
Task-number: QTBUG-42012
Reviewed-by: Christian Stromme <christian.stromme@digia.com>
Reviewed-by: BogDan Vatra <bogdan@kde.org>