macdeployqt: Also deploy framework's Libraries and Helpers
QtWebEngine can't put all its external data in qrc files and have a few
files lying around in the Qt installation. To allow it to be bundled
with applications, those files are now placed into the QtWebEngineCore
framework.

This patch allows Helpers/QtWebEngineProcess.app as well as
Libraries/ffmpegsumo.so to be deployed when macdeployqt is run on an
application linking to QtWebEngine.

This link mention that Libraries is a standard directory:
https://developer.apple.com/library/mac/documentation/MacOSX/Conceptual/BPFrameworks/Concepts/FrameworkAnatomy.html

This link mention that Helpers is a standard directory for application
bundles, but I haven't seen any mention regarding frameworks:
https://developer.apple.com/library/mac/technotes/tn2206/_index.html

Task-number: QTBUG-41611
Change-Id: I27f662747bd3dff4b9fc8706ecbd2f16f057e857
(cherry picked from commit 5984871db094182b89d2b8e0b142817b227afa4d)
Reviewed-by: Morten Johan Sørvig <morten.sorvig@digia.com>