QHelpContentModel: Simplify content update
Since an update always replaces the whole model, we can use begin/
endResetModel instead of removing/adding ranges of rows.
Also, the use of d->qhelpContentProvider->nextChildCount() was wrong
anyhow, because newRootItem = d->qhelpContentProvider->rootItem()
removes the root item from the list, so when nextChildCount() is called
later, it does _not_ return the child count of the retrieved
newRootItem.
(Aside from that nextChildCount() was missing a lock, but the method is
not needed anymore anyhow.)

Change-Id: Iafad0a43ad4c33075c327926c78d53ba1c310b13
Reviewed-by: Christian Kandeler <christian.kandeler@theqtcompany.com>
Reviewed-by: Karsten Heimrich <karsten.heimrich@theqtcompany.com>