Fix build after commit 9d7b27f5bf44a46707e6d50ebf51ecf73f91dd1b in qtdeclarative
The commit integrates the .pragma and .import parsing in JavaScript files
directly into the lexer. Previously the QML engine had a separate function
for parsing those directives and replacing them with whitespace in the source.
Similarly the code here tried to do the same. Fortunately this is not necessary
anymore as the lexer now understands those directives and unless told otherwise
will simply ignore them.

Change-Id: If9888deb2c82a15d1727f4254d897fc9e7bcb832
Reviewed-by: Oswald Buddenhagen <oswald.buddenhagen@theqtcompany.com>
Reviewed-by: Simon Hausmann <simon.hausmann@digia.com>