Change creation date to a fixed date
Encoding the dates makes reproducing the output impossible, and doesn't
really have much gain. It most of the time not relevant when something
was built, but more 'what was built' and 'with what tools was it built'

In this case, the creation date isn't read by any thing, so it is likely
just filler data.

Keep the key anyways to not break any third party readers of the file
and set a fixed date, in this case the same date as used for a fixed
date in QtCore's QLibraryInfo::buildDate function at noon.

Change-Id: I560e43dfce538e2f25c3cf51482b946a7cfa179e
Reviewed-by: Christian Kandeler <christian.kandeler@theqtcompany.com>