Qt Designer: Enable reset property for multiselection.
When working with a multiselection in the property editor, the reset
button was previously only enabled when the property was modified on
the selected/current widget. With this change, reset will always be
enabled when the property is modified in one of the widgets of
a multiselection. Unmodified widgets will be filtered out when
initializing the command.

The intended use case is mainly to enable conveniently clearing all
minimumSize/maximumSize constraints of a form by doing Ctrl+A
to select all and then do reset on the property.

Change-Id: Ideab87cab79e9404941d5594653f12c13515c1ae
Reviewed-by: Jarek Kobus <jaroslaw.kobus@theqtcompany.com>