move ui loader interfaces to own module
it's backwards to declare the interface as part of the designer library
itself, as that obviously pulls in half of designer itself into custom
widget plugins which should be perfectly usable from the stand-alone
ui loader in uitools. also, designer isn't built for embedded systems,
so the bundling needlessly limited the availability of widget plugins.

so move the relevant headers into a separate include-only module.
the designer module retains forwarding headers, as the old includes
obviously need to continue working.

[ChangeLog][UiTools] Custom widget plugins should now use QT+=uiplugin
instead of QT+=designer. This makes them usable on embedded systems.

Task-number: QTBUG-44724
Change-Id: I5535b10d2f095b257eeda9681f60b2f8c699f7d7
Reviewed-by: Friedemann Kleint <Friedemann.Kleint@theqtcompany.com>