update message index when message gains source text
when a message which is first seen with only an id later gains a source
text, this text would not have been indexed. so a subsequent new message
with the same source, but without an id this time around, would not be
seen as a dupe of the first message.

Change-Id: Ib35f820e2b02311ea9c9996383a2789095098d6f
Reviewed-by: hjk <hjk@theqtcompany.com>