windeployqt: Patch qt_prfxpath when deploying Qt5Core
Patch the preconfigured Qt path in Qt5Core to '.'. This allows one to
immediately test whether the generated package is self-contained, without
Qt being able to 'fall back' to the installed Qt for loading plugins. It
is also fixing a potential security issue where Qt tries to load plugins
from 'unexpected' places.

The effect is the same as adding a qt.conf file with all the defaults.
However, a lot of people consider adding a qt.conf ugly.

[ChangeLog][Important behavior changes][windeployqt] windeployqt now
changes the hardcoded QLibraryInfo::PrefixPath in Qt5Core to be relative
to the executable.

Change-Id: Ib59d84453913f28b806ae78a0ef138eb4f0045fa
Reviewed-by: Oliver Wolff <oliver.wolff@theqtcompany.com>
Reviewed-by: Björn Breitmeyer <bjoern.breitmeyer@kdab.com>
Reviewed-by: Lars Knoll <lars.knoll@theqtcompany.com>
Reviewed-by: Friedemann Kleint <Friedemann.Kleint@theqtcompany.com>