windeployqt: Do not patch Qt5Core on WinRT/Windows Phone
Patching the file causes the "Date modified" to be changed. When trying
to deploy a package using Visual Studio, MSVC does not like the
date/time difference between the file from the package and Qt's bin
directory and thus refuses to deploy the package.

As WinRT/Windows phone packages are sandboxed and have to be self-
contained patching Qt5Core is not that important on these platforms and
can be omitted.

Change-Id: I03404976efa0f343341b5d13708ebb864a3b8179
Reviewed-by: Andrew Knight <andrew.knight@intopalo.com>
Reviewed-by: Friedemann Kleint <Friedemann.Kleint@theqtcompany.com>