windeployqt: Fix the module name matching to longer rely on substring search.
Previously, module name matching worked by determining the best match
length, which caused Qt5QuickTemplates to be reported as Qt5Quick.
Rewrite it to strip down the needle module name so that it can be matched
directly, without the need to check for a best match. The libinfix is
required for this as a parameter, which is extracted from the qconfig.pri
written by configure.

Task-number: QTBUG-48386
Change-Id: Ie05b32fc1785a3a770e0c4aed533866f7fdf512f
Reviewed-by: J-P Nurmi <jpnurmi@theqtcompany.com>
Reviewed-by: Oliver Wolff <oliver.wolff@theqtcompany.com>