qdoc: Remove creation of redundant function & variable nodes
qdoc no longer creates a function node when it parses a function
declaration, if a function node for the function already exists.
This avoids needless allocation and deallocation.

qdoc now also avoids creating unnecessary variable nodes.

This change also improves the "No documentation for..." error
message so that it includes the parameters if the error refers
to a function. This makes it easier to find the correct function
when you are trying to correct these errors in the documentation.

Change-Id: Ib0f9361622f9115d677408e36bfed77c42545be0
Task-number: QTBUG-48191
Reviewed-by: Martin Smith <martin.smith@digia.com>