Doc: Generate correct TOC for Qt Linguist Manual
The Qt Linguist documentation structure follows the rules for a QHP
manual, that is, the landing page contains a \list with links to
all the pages in the manual.

Mark the Qt Help section as a type 'manual'; this generates a
better-looking TOC without a duplicate nested entry for 'Manual'.

Change-Id: Id59c645a9a9fb945ddf8999cd0952c81059ac5b9
Task-number: QTBUG-32989
Reviewed-by: Martin Smith <martin.smith@digia.com>