qdoc: Fix a potential crash in generation of Table of Contents
There was nonsensical code in HtmlGenerator::generateTableOfContents(),
it could be simplified a lot without changing the functionality and
avoid a potential crash.

Also, related to ToC, marked the correct header offsets for module and
QML module pages. The former gets an automatic 'Detailed Description'
added to ToC, the latter does not. Previously, both were assigned a
default offset of 3, meaning that \section1 headers on those pages were
generated incorrectly as <h4> sub-sub-subheaders.

Change-Id: I425ed9d5a333bc8959ec2bb1d255214cf2b45ec0
Reviewed-by: Martin Smith <martin.smith@theqtcompany.com>