qdoc: Add support for listing all examples from all modules
This commit adds a new argument for \generatelist command, which
enables the generation of an 'example index' page:

    \generatelist annotatedexamples

This will generate a set of tables; one for each module, prefixed
by a header titled with the module name. The table is an annotated
list (title + brief description) for each documented example.

A human-readable name for each module is read from the
navigation.landingpage .qdocconf variable. This name is written
to / read back from the index files as 'indexTitle' attribute.
Examples for modules that do not define a landing page title
are listed in a single table with no accompanying header.

Change-Id: Icb7fa6fb40ec9a0dfce1bf8282c8782a69ec0028
Reviewed-by: Venugopal Shivashankar <venugopal.shivashankar@digia.com>
Reviewed-by: Martin Smith <martin.smith@theqtcompany.com>