qdoc: sub-class command-line parser to declutter main()
More than two thirds of main() concerned parsing the command-line
arguments and settings globals, interacting minimally with the rest.
So split that code out into a sub-class of the command-line parser, to
make main() itself easier to read.

Changed a QStringLiteral to a QLatin1String to work round an MSVC-2010
issue.

Change-Id: Idd5fec280b0ce631f1d6d2a79cb59f9c4171eceb
Reviewed-by: Topi Reiniö <topi.reinio@theqtcompany.com>