lupdate: Support for C++11's "final" identifier
C++11 introduced the "final" identifier which may appear immediately
after the name in a class definition to specify that the class cannot be
derived from.

lupdate collects strings for translation and did not understand "final".
Instead it would use "final" as the name of the class as there may be
attributes before the class name and the last token before the colon
character is used.

[ChangeLog][lupdate] The C++11 identifier "final" is now skipped when
determining the name of a class.

Task-number: QTBUG-42735
Change-Id: I2adee9a188645622fc6916158ba2b48dde921383
Reviewed-by: Oswald Buddenhagen <oswald.buddenhagen@theqtcompany.com>