lupdate: Merge class name extraction loops
The extraction of class names was split over two separate loops. One for
plain identifiers (e.g. “class Foo”) and another for qualified
identifiers (e.g. “class Foo::Bar::Baz”).

C++11 introduced the "final" identifier which may appear immediately
after the name in a class definition. Detecting and skipping that
identifier will only need to be in a single place when the two loops are
merged.

Change-Id: Id9ddc1ccd01cabcb4616d8c025b553c5be5c5807
Reviewed-by: Oswald Buddenhagen <oswald.buddenhagen@theqtcompany.com>