qdoc: Don't let a QML type be its own base type
When qdoc runs without one or more of the index files for
the modules it depends on, there was a chance it could set
a QML type's base type to point to the QML type itself.
This resulted in an infinite loop, which crashed qdoc with
a bad alloc error. qdoc has now been changed so that it
refuses to set a QML type's base type to itself.

Change-Id: I08e8959ddb67b2d1f3b1bf44d9dc48624bafebfa
Task-number: QTBUG-50163
Reviewed-by: Martin Smith <martin.smith@theqtcompany.com>