Highlight QTabWidget tabs by making them current
If the text to translate is a QTabWidget tab label, the whole QTabWidget is
highlighted but the tab inconspicuously remains unselected. The translator
wastes precious seconds searching for the text in the preview (at least I
do, even with three tabs).

Moreover, automatically bringing the tab to the front let the translator
see what the (usually terse) tab label refers to.

Change-Id: I85ee872403b276ebd740cfb9542efc1f7debb40c
Reviewed-by: Oswald Buddenhagen <oswald.buddenhagen@theqtcompany.com>