Fix macdeployqt to properly sign inner bundles.
QtWebEngine uses a framework called QtWebEngineCore which contains
an inner bundle called QtWebEngineProcess.app. Because it was not
signed in the proper order, the whole signing process failed.

Fix consists in checking if there are any inner app bundles
inside frameworks, and sign them before signing the rest of the
main bundle.

Change-Id: I48abe26d1e61cd1ce17cb185bcd8d1d72f6c9607
Task-number: QTBUG-50636
Reviewed-by: Morten Johan Sørvig <morten.sorvig@theqtcompany.com>