qdoc: Don't print unnecessary qdoc warnings
Generates default doc for destructor that is declared but not
documented, thereby avoiding the "No documentation for..." error.

This change reveals several new constructors and assignment operators
that are not documented. Most of these will be handled automatically
the same way these destructors are handled, but that fix will be in a
separate update.

Change-Id: Iffb3834b2ea31fd47d578bf1444be2f25a60ae6f
Task-number: QTBUG-50630
Reviewed-by: Topi Reiniö <topi.reinio@theqtcompany.com>