macdeployqt: Fix code-signing errors.
Starting from Qt 5.6 all frameworks / plugins contain an embedded
rpath. macdeployqt reads the rpath to get the dependencies which
should be signed first, and because the rpath points to a "lib"
directory which does not exist in the bundle structure, it shows
errors. The solution is to additionally use the rpaths embedded
in the bundle executable.

Also because on 5.5 and previously rpaths were not included in
frameworks, the getBinaryDependencies() always returned an
empty set of dependencies, because it didn't have an rpath
to resolve the paths to the dependencies (which was probably a
bug). Now that the rpaths are present, the dependency list for
a framework also contains itself as a dependency, which leads
to an infinite loop while code-signing. The fix consists in not
repeatedly adding binaries that were alread scheduled for
code-singning.

Change-Id: I97b418b90f325f88f375ffec83ad68c7dd4528ca
Task-number: QTBUG-51101
Reviewed-by: Morten Johan Sørvig <morten.sorvig@theqtcompany.com>