qdoc: Ensure all aggregate nodes are processed and added to QHP
When recursively traversing the tree of aggregates, QDoc previously
set them in a QMap with fullDocumentName() as a key. Aggregates can
(and often do) have identical names, for example, QtConcurrent (the
module) and QtConcurrent (the namespace). This led to only one of
them being added to QHP.

Fix the issue by replacing the QMap with a QSet for storing the
aggregate's child nodes.

Change-Id: I5107c057b349b1c1c2b1ed505f1ace99b6f96d9a
Task-number: QTBUG-51358
Reviewed-by: Martin Smith <martin.smith@theqtcompany.com>