qdoc: Fix link disambiguation
The link command allows a qualifier in square brackets,
which tells qdoc the specific kind of page that must be
found to satisfy the link. The square bracket parameter
was being ignored in one case, because it wasn't being
passed to a certain search function. This update passes
that parameter to the search function, and the search
function uses it to disqualify pages that should not be
accepted.

This fix causes qdoc to add 4 new qdoc errors when
generating the docs for my Qt5 installation. I don't
know what these are, but I suspect they are real errors
that qdoc was not reporting because the square bracket
parameter was being ignored in this case.

Change-Id: I45f1fd43039df7bb048711fc5b752a71a006e5de
Task-number: QTBUG-51587
Reviewed-by: Topi Reiniö <topi.reinio@theqtcompany.com>