qdoc: Provide better threadsafeness reporting
The documentation QObject reported the thread-safety of the
class in a kind of silly way. It said "Note: All functions in
this class are rentrant except for connect(), connect(),
connect(), ..." This was confusing to the reader. Now it says
"Note: All functions in this class are reentrant." And then
it shows a second note that lists all the functions that are
thread-safe in addition to being reentrant, but it lists them
in a bullet list with their complete signatures, except for
the return type. This affects a few other classes that had a
similar problem.

Change-Id: Icb7f7c5ac2da1c28ef8e5c10736296d80c2aefeb
Task-number: QTBUG-41724
Reviewed-by: Topi Reiniö <topi.reinio@theqtcompany.com>