Parse function pointer types the same way clang does
When an argument to a function is itself a function, use
the same representation of its type that clang uses.
 - Omit spaces adjacent to parentheses
 - The argument name is not part of the argument type.

This is required so types parsed with our parser lead to the same
string as libclang gives us.

The only function affected in QtCore is qRegisterAnimationInterpolator.
Before: void qRegisterAnimationInterpolator(QVariant(* ) ( const T & from, const T & to, qreal progress ) func)
After: void qRegisterAnimationInterpolator(QVariant (*)(const T &, const T &, qreal) func)

Unfortunately, this means the documentation does not contain the parameter
names in the function type. We will probably have to adapt the documentation
of qRegisterAnimationInterpolator() to mention the meaning explicitly.

Change-Id: I00ba1854869c061fb54c6a53e6eb02b934c20459
Reviewed-by: Topi Reiniö <topi.reinio@qt.io>