qdoc: Use '//!' as a documentation comment
This commit makes QDoc treat a string starting with '//!' as a
single-line comment that is not included in the generated HTML
output.

It has several uses: Internal documentation remarks visible only in
the source code, and a simpler alternative to \omit, \endomit commands.
Most importantly, it allows \include command to pull in content from
any file mentioned in the source directories, for example:

  /*!
    \fn foo()

    //! [foo-descr]
    ...long description...
    //! [foo-descr]
  */

  /*/
    \fn foo(int bar)
    \overload foo()
    \include foo.cpp foo-descr
    ...
  */

Above will work and the snippet identifies are omitted from the
generated HTML output.

Change-Id: I0777247dc00358682c307c356fb7d1d8a4514dac
Task-number: QTBUG-52063
Reviewed-by: Martin Smith <martin.smith@theqtcompany.com>