Fix macdeployqt usage when client application has no rpath to Qt libs.
macdeployqt used to determine the Qt library path from one of the Qt
libraries linked to by the application (by searching its rpath list).
However, if an application does not have an rpath pointing to the Qt
library directory, macdeployqt cannot find the Qt libraries. This is
likely when the application is built with a non-qmake build system like
CMake or qbs.

This patch adds the Qt library path (from QLibraryInfo) as a fallback
search path to allow macdeployqt to find the Qt libraries and deploy
them.

Task-number: QTBUG-47868
Change-Id: Ie1d005955826b416c223a1bc6cac911d2ae44b20
Reviewed-by: Guilherme Steinmann <guidefloripa@gmail.com>
Reviewed-by: Richard Moe Gustavsen <richard.gustavsen@theqtcompany.com>